+++ lib/transformers.dart
@@ -1,5 +1,5 @@
-import 'dart:typed_data';
 import 'dart:async';
+import 'dart:typed_data';
 
 /// wildcardFind searches the haystack for a copy of needle.
 /// both needle and haystack must be Lists. Needle may
@@ -31,14 +31,18 @@ int wildcardFind(dynamic needle, dynamic haystack) {
   return -1;
 }
 
+/// parent abstract class to all transformers adding dispose method.
+abstract class DisposableStreamTransformer<T, R> implements StreamTransformer<T, R> {
+  void dispose();
+}
+
 /// This transformer takes an incoming stream and splits it
 /// along the "terminator" marks. Great for parsing incoming
 /// serial streams that end on \r\n.
-class TerminatedTransformer implements StreamTransformer<Uint8List, Uint8List> {
+class TerminatedTransformer implements DisposableStreamTransformer<Uint8List, Uint8List> {
   bool? cancelOnError;
   final Uint8List? terminator;
-  final int
-      maxLen; // maximum length the partial buffer will hold before it starts flushing.
+  final int maxLen; // maximum length the partial buffer will hold before it starts flushing.
   final stripTerminator;
 
   late StreamController _controller;
@@ -55,12 +59,7 @@ class TerminatedTransformer implements StreamTransformer<Uint8List, Uint8List> {
   /// remain attached in the output.
   ///
   /// This constructor creates a single stream
-  TerminatedTransformer(
-      {bool sync: false,
-      this.cancelOnError,
-      this.terminator,
-      this.maxLen = 1024,
-      this.stripTerminator = true}) {
+  TerminatedTransformer({bool sync: false, this.cancelOnError, this.terminator, this.maxLen = 1024, this.stripTerminator = true}) {
     _partial = [];
     _controller = new StreamController<Uint8List>(
         onListen: _onListen,
@@ -83,22 +82,13 @@ class TerminatedTransformer implements StreamTransformer<Uint8List, Uint8List> {
   /// remain attached in the output.
   ///
   /// This constructor creates a broadcast stream
-  TerminatedTransformer.broadcast(
-      {bool sync: false,
-      this.cancelOnError,
-      this.terminator,
-      this.maxLen = 1024,
-      this.stripTerminator = true}) {
+  TerminatedTransformer.broadcast({bool sync: false, this.cancelOnError, this.terminator, this.maxLen = 1024, this.stripTerminator = true}) {
     _partial = [];
-    _controller = new StreamController<Uint8List>.broadcast(
-        onListen: _onListen, onCancel: _onCancel, sync: sync);
+    _controller = new StreamController<Uint8List>.broadcast(onListen: _onListen, onCancel: _onCancel, sync: sync);
   }
 
   void _onListen() {
-    _subscription = _stream.listen(onData,
-        onError: _controller.addError,
-        onDone: _controller.close,
-        cancelOnError: cancelOnError);
+    _subscription = _stream.listen(onData, onError: _controller.addError, onDone: _controller.close, cancelOnError: cancelOnError);
   }
 
   void _onCancel() {
@@ -121,9 +111,9 @@ class TerminatedTransformer implements StreamTransformer<Uint8List, Uint8List> {
       if (stripTerminator) {
         message = Uint8List.fromList(_partial.take(index).toList());
       } else {
-        message = Uint8List.fromList(
-            _partial.take(index + terminator!.length).toList());
+        message = Uint8List.fromList(_partial.take(index + terminator!.length).toList());
       }
+
       _controller.add(message);
       _partial = _partial.sublist(index + terminator!.length);
     }
@@ -136,19 +126,21 @@ class TerminatedTransformer implements StreamTransformer<Uint8List, Uint8List> {
   }
 
   @override
-  StreamTransformer<RS, RT> cast<RS, RT>() =>
-      StreamTransformer.castFrom<Uint8List, Uint8List, RS, RT>(this);
+  StreamTransformer<RS, RT> cast<RS, RT>() => StreamTransformer.castFrom<Uint8List, Uint8List, RS, RT>(this);
+
+  @override
+  void dispose() {
+    _controller.close();
+  }
 }
 
 /// This transformer takes an incoming stream and splits it
 /// along the "terminator" marks and returns a String! Great
 /// for parsing incoming serial streams that end on \r\n.
-class TerminatedStringTransformer
-    implements StreamTransformer<Uint8List, String> {
+class TerminatedStringTransformer implements DisposableStreamTransformer<Uint8List, String> {
   bool? cancelOnError;
   final Uint8List? terminator;
-  final int
-      maxLen; // maximum length the partial buffer will hold before it starts flushing.
+  final int maxLen; // maximum length the partial buffer will hold before it starts flushing.
   final stripTerminator;
 
   late StreamController _controller;
@@ -167,12 +159,7 @@ class TerminatedStringTransformer
   /// remain attached in the output.
   ///
   /// This constructor creates a single string stream
-  TerminatedStringTransformer(
-      {bool sync: false,
-      this.cancelOnError,
-      this.terminator,
-      this.maxLen = 1024,
-      this.stripTerminator = true}) {
+  TerminatedStringTransformer({bool sync: false, this.cancelOnError, this.terminator, this.maxLen = 1024, this.stripTerminator = true}) {
     _partial = [];
     _controller = new StreamController<String>(
         onListen: _onListen,
@@ -197,22 +184,13 @@ class TerminatedStringTransformer
   /// remain attached in the output.
   ///
   /// This constructor creates a broadcast string stream
-  TerminatedStringTransformer.broadcast(
-      {bool sync: false,
-      this.cancelOnError,
-      this.terminator,
-      this.maxLen = 1024,
-      this.stripTerminator = true}) {
+  TerminatedStringTransformer.broadcast({bool sync: false, this.cancelOnError, this.terminator, this.maxLen = 1024, this.stripTerminator = true}) {
     _partial = [];
-    _controller = new StreamController<String>.broadcast(
-        onListen: _onListen, onCancel: _onCancel, sync: sync);
+    _controller = new StreamController<String>.broadcast(onListen: _onListen, onCancel: _onCancel, sync: sync);
   }
 
   void _onListen() {
-    _subscription = _stream.listen(onData,
-        onError: _controller.addError,
-        onDone: _controller.close,
-        cancelOnError: cancelOnError);
+    _subscription = _stream.listen(onData, onError: _controller.addError, onDone: _controller.close, cancelOnError: cancelOnError);
   }
 
   void _onCancel() {
@@ -235,8 +213,7 @@ class TerminatedStringTransformer
       if (stripTerminator) {
         message = Uint8List.fromList(_partial.take(index).toList());
       } else {
-        message = Uint8List.fromList(
-            _partial.take(index + terminator!.length).toList());
+        message = Uint8List.fromList(_partial.take(index + terminator!.length).toList());
       }
       _controller.add(String.fromCharCodes(message));
       _partial = _partial.sublist(index + terminator!.length);
@@ -250,8 +227,12 @@ class TerminatedStringTransformer
   }
 
   @override
-  StreamTransformer<RS, RT> cast<RS, RT>() =>
-      StreamTransformer.castFrom<Uint8List, String, RS, RT>(this);
+  StreamTransformer<RS, RT> cast<RS, RT>() => StreamTransformer.castFrom<Uint8List, String, RS, RT>(this);
+
+  @override
+  void dispose() {
+    _controller.close();
+  }
 }
 
 /// Assembles messages that start with a fixed sequence of zero or more magic
@@ -267,27 +248,20 @@ class TerminatedStringTransformer
 /// 0x10 0x04 0x01 0x02 0x03 0x04
 ///
 /// Will clear input if no data is received for at least 1 second.
-class MagicHeaderAndLengthByteTransformer
-    implements StreamTransformer<Uint8List, Uint8List> {
+class MagicHeaderAndLengthByteTransformer implements DisposableStreamTransformer<Uint8List, Uint8List> {
   final List<int?>? header;
   final Duration clearTimeout;
   late List<int> _partial;
   Timer? _timer;
   late bool _dataSinceLastTick;
   bool? cancelOnError;
-  final int
-      maxLen; // maximum length the partial buffer will hold before it starts flushing.
+  final int maxLen; // maximum length the partial buffer will hold before it starts flushing.
 
   late StreamController _controller;
   StreamSubscription? _subscription;
   late Stream<Uint8List> _stream;
 
-  MagicHeaderAndLengthByteTransformer(
-      {bool sync: false,
-      this.cancelOnError,
-      this.header,
-      this.maxLen = 1024,
-      this.clearTimeout = const Duration(seconds: 1)}) {
+  MagicHeaderAndLengthByteTransformer({bool sync: false, this.cancelOnError, this.header, this.maxLen = 1024, this.clearTimeout = const Duration(seconds: 1)}) {
     _partial = [];
     _controller = new StreamController<Uint8List>(
         onListen: _onListen,
@@ -302,22 +276,14 @@ class MagicHeaderAndLengthByteTransformer
   }
 
   MagicHeaderAndLengthByteTransformer.broadcast(
-      {bool sync: false,
-      this.cancelOnError,
-      this.header,
-      this.maxLen = 1024,
-      this.clearTimeout = const Duration(seconds: 1)}) {
+      {bool sync: false, this.cancelOnError, this.header, this.maxLen = 1024, this.clearTimeout = const Duration(seconds: 1)}) {
     _partial = [];
-    _controller = new StreamController<Uint8List>.broadcast(
-        onListen: _onListen, onCancel: _onCancel, sync: sync);
+    _controller = new StreamController<Uint8List>.broadcast(onListen: _onListen, onCancel: _onCancel, sync: sync);
   }
 
   void _onListen() {
     _startTimer();
-    _subscription = _stream.listen(onData,
-        onError: _controller.addError,
-        onDone: _controller.close,
-        cancelOnError: cancelOnError);
+    _subscription = _stream.listen(onData, onError: _controller.addError, onDone: _controller.close, cancelOnError: cancelOnError);
   }
 
   void _onCancel() {
@@ -355,8 +321,7 @@ class MagicHeaderAndLengthByteTransformer
         return;
       }
 
-      _controller.add(
-          Uint8List.fromList(_partial.sublist(0, len + header!.length + 1)));
+      _controller.add(Uint8List.fromList(_partial.sublist(0, len + header!.length + 1)));
       _partial = _partial.sublist(len + header!.length + 1);
     }
   }
@@ -368,8 +333,7 @@ class MagicHeaderAndLengthByteTransformer
   }
 
   @override
-  StreamTransformer<RS, RT> cast<RS, RT>() =>
-      StreamTransformer.castFrom<Uint8List, Uint8List, RS, RT>(this);
+  StreamTransformer<RS, RT> cast<RS, RT>() => StreamTransformer.castFrom<Uint8List, Uint8List, RS, RT>(this);
 
   void _onTimer(Timer timer) {
     if (_partial.length > 0 && !_dataSinceLastTick) {
@@ -386,5 +350,10 @@ class MagicHeaderAndLengthByteTransformer
   void _startTimer() {
     _dataSinceLastTick = false;
     _timer = Timer.periodic(clearTimeout, this._onTimer);
+  }
+
+  @override
+  void dispose() {
+    _controller.close();
   }
 }