+++ CHANGELOG.md
@@ -1,3 +1,9 @@
+## 0.3.0
+
+* Implemented v2 Android embedding
+* Updated UsbSerial to 6.1.0
+* Transaction constructor signature changed to fix issue #39 (see test testTwoTransformersInSequence)
+
 ## 0.2.991
 
 * Addressed issue #46